Creates input package and renames inputReader to reader
