--- src/ttt/cli/cliinputreader.go
@@ -1,14 +1,0 @@
-package cli
-
-import (
-	"bufio"
-)
-
-type CliInputReader struct {
-	reader *bufio.Reader
-}
-
-func (input CliInputReader) read() string {
-	userInput, _ := input.reader.ReadString('\n')
-	return userInput
-}
