+++ src/ttt/cli/inputreader.go
@@ -1,5 +1,5 @@
 package cli
 
 type InputReader interface {
-	read() string
+	Read() string
 }
